+++ 1.Domain/Domain/Repositories/IMoviePlayRepository.cs
@@ -1,5 +1,5 @@
 ﻿using Domain.Entities;
-using Domain.Repositories.BaseReposotries;
+using Domain.Repositories.BaseRepositories;
 
 namespace Domain.Repositories;
 public interface IMoviePlayRepository : IRepository<MoviePlay>
